requirements.txt
langchain
langchain-community
llama-cpp-python
openai
selenium
Pillow
webdriver-manager